+++ CGProj/Source/Utils/MeshImporter.cpp
@@ -11,7 +11,7 @@
 #include "Core/Objects/Mesh.h"
 
 
-Mesh* MeshImporter::ImportMesh(const std::string& pathModel, const wchar_t* pathTex, bool centering, int32_t numMesh)
+Mesh* MeshImporter::ImportMesh(const std::string& pathModel, const wchar_t* pathTex, bool centering, uint32_t numMesh)
 {
     Assimp::Importer importer;
     const aiScene* sce = importer.ReadFile(pathModel, aiProcess_Triangulate | aiProcess_GenBoundingBoxes);
@@ -25,7 +25,7 @@ Mesh* MeshImporter::ImportMesh(const std::string& pathModel, const wchar_t* path
     return resMesh;
 }
 
-bool MeshImporter::ImportMesh(const std::string& pathModel, Mesh* resMesh, bool centering, int32_t numMesh)
+bool MeshImporter::ImportMesh(const std::string& pathModel, Mesh* resMesh, bool centering, uint32_t numMesh)
 {
     Assimp::Importer importer;
     const aiScene* sce = importer.ReadFile(pathModel, aiProcess_Triangulate | aiProcess_GenBoundingBoxes);
@@ -34,15 +34,15 @@ bool MeshImporter::ImportMesh(const std::string& pathModel, Mesh* resMesh, bool 
     return true;
 }
 
-void MeshImporter::FillMeshData(const aiScene* sce, Mesh* resMesh, bool centering, int32_t numMesh)
+void MeshImporter::FillMeshData(const aiScene* sce, Mesh* resMesh, bool centering, uint32_t numMesh)
 {
     if (!resMesh) return;
 
     int32_t numVertices = 0;
-    for (int32_t i = 0; i < (sce->mNumMeshes && numMesh && numMesh <= sce->mNumMeshes ? numMesh : sce->mNumMeshes); i++)
+    for (uint32_t i = 0; i < (sce->mNumMeshes && numMesh && numMesh <= sce->mNumMeshes ? numMesh : sce->mNumMeshes); i++)
     {
         const auto mesh = sce->mMeshes[i];
-        for (int32_t j = 0; j < mesh->mNumVertices; j++)
+        for (uint32_t j = 0; j < mesh->mNumVertices; j++)
             numVertices++;
     }
     int32_t sideSize = (int32_t)std::sqrt(numVertices) + 1;
@@ -51,19 +51,19 @@ void MeshImporter::FillMeshData(const aiScene* sce, Mesh* resMesh, bool centerin
     Vector3 maxAABB = Vector3::Zero;
     Vector3 minAABB = Vector3::Zero;
     Array<Vertex> importingVert;
-    for (int32_t i = 0; i < (sce->mNumMeshes && numMesh && numMesh <= sce->mNumMeshes ? numMesh : sce->mNumMeshes); i++)
+    for (uint32_t i = 0; i < (sce->mNumMeshes && numMesh && numMesh <= sce->mNumMeshes ? numMesh : sce->mNumMeshes); i++)
     {
         const auto mesh = sce->mMeshes[i];
         printf("Vertex %d\n", mesh->mNumVertices);
         Vector3::Max(maxAABB, Vector3(mesh->mAABB.mMax.x, mesh->mAABB.mMax.y, mesh->mAABB.mMax.z), maxAABB);
         Vector3::Min(minAABB, Vector3(mesh->mAABB.mMin.x, mesh->mAABB.mMin.y, mesh->mAABB.mMin.z), minAABB);
-        for (int32_t j = 0; j < mesh->mNumVertices; j++)
+        for (uint32_t j = 0; j < mesh->mNumVertices; j++)
         {
             Vertex newVert = {Vector4(mesh->mVertices[j].x, mesh->mVertices[j].y, mesh->mVertices[j].z, 1.0f),
                               Vector4(mesh->mNormals[j].x, mesh->mNormals[j].y, mesh->mNormals[j].z, 0.0f),
                               mesh->HasTextureCoords(0)
                                   ? Vector2(mesh->mTextureCoords[0][j].x, mesh->mTextureCoords[0][j].y)
-                                  : Vector2((float)((count + j) % sideSize) / sideSize, (float)((count + j) / sideSize) / sideSize)
+                                  : Vector2((float)((count + j) % sideSize) / (float)sideSize, (float)((count + j) / sideSize) /(float)sideSize)
             };
             /*printf("x = %f, y = %f\n", newVert.texture.x,
                 newVert.texture.y);*/
@@ -73,16 +73,16 @@ void MeshImporter::FillMeshData(const aiScene* sce, Mesh* resMesh, bool centerin
 
         count += mesh->mNumVertices;
     }
-    for (int32_t i = 0; i < (sce->mNumMeshes && numMesh && numMesh <= sce->mNumMeshes  ? numMesh : sce->mNumMeshes); i++)
+    for (uint32_t i = 0; i < (sce->mNumMeshes && numMesh && numMesh <= sce->mNumMeshes  ? numMesh : sce->mNumMeshes); i++)
     {
         const auto mesh = sce->mMeshes[i];
         printf("Vertex %d\n", mesh->mNumVertices);
         Vector3::Max(maxAABB, Vector3(mesh->mAABB.mMax.x, mesh->mAABB.mMax.y, mesh->mAABB.mMax.z), maxAABB);
         Vector3::Min(minAABB, Vector3(mesh->mAABB.mMin.x, mesh->mAABB.mMin.y, mesh->mAABB.mMin.z), minAABB);
-        for (int32_t j = 0; j < mesh->mNumFaces; j++)
+        for (uint32_t j = 0; j < mesh->mNumFaces; j++)
         {
             aiFace face = mesh->mFaces[j];
-            for (int32_t k = 0; k < face.mNumIndices; k++)
+            for (uint32_t k = 0; k < face.mNumIndices; k++)
             {
                 resMesh->AddIndex(face.mIndices[k]);
             }
